Fix hardware invoke + refactor hardware + refactor tests, constants
