+++ src-tauri/plugins/tauri-plugin-hardware/permissions/default.toml
@@ -1,12 +1,6 @@
-# Default permission for the tauri-plugin-hardware plugin
-# This permission allows access to all hardware plugin commands
-
-"$schema" = "./schemas/schema.json"
-
-[[permission]]
-identifier = "default"
-description = "enables the default permissions for the hardware plugin"
+[default]
+description = "Default permissions for the hardware plugin"
 permissions = [
-  "allow-get-system-info",
-  "allow-get-system-usage"
+    "allow-get-system-info",
+    "allow-get-system-usage"
 ]
