+++ src-tauri/plugins/tauri-plugin-hardware/src/constants.rs
@@ -1,0 +1,4 @@
+// https://devicehunt.com/all-pci-vendors
+pub const VENDOR_ID_AMD: u32 = 0x1002;
+pub const VENDOR_ID_NVIDIA: u32 = 0x10DE;
+pub const VENDOR_ID_INTEL: u32 = 0x8086;
