+++ src-tauri/src/core/app/constants.rs
@@ -1,0 +1,2 @@
+// App Configuration Constants
+pub const CONFIGURATION_FILE_NAME: &str = "settings.json";
