+++ src-tauri/src/core/app/mod.rs
@@ -1,3 +1,4 @@
 pub mod commands;
+mod constants;
 pub mod helpers;
 pub mod models;
