+++ src-tauri/src/core/mcp/mod.rs
@@ -1,4 +1,5 @@
 pub mod commands;
+mod constants;
 pub mod helpers;
 
 #[cfg(test)]
