+++ src-tauri/src/core/state.rs
@@ -6,7 +6,6 @@ use tokio::task::JoinHandle;
 
 /// Server handle type for managing the proxy server lifecycle
 pub type ServerHandle = JoinHandle<Result<(), Box<dyn std::error::Error + Send + Sync>>>;
-use tauri_plugin_llamacpp::LLamaBackendSession;
 use tokio::sync::Mutex;
 
 #[derive(Default)]
@@ -18,5 +17,4 @@ pub struct AppState {
     pub mcp_active_servers: Arc<Mutex<HashMap<String, serde_json::Value>>>,
     pub mcp_successfully_connected: Arc<Mutex<HashMap<String, bool>>>,
     pub server_handle: Arc<Mutex<Option<ServerHandle>>>,
-    pub llama_server_process: Arc<Mutex<HashMap<i32, LLamaBackendSession>>>,
 }
