+++ src-tauri/src/core/threads/constants.rs
@@ -1,0 +1,4 @@
+// Thread Constants
+pub const THREADS_DIR: &str = "threads";
+pub const THREADS_FILE: &str = "thread.json";
+pub const MESSAGES_FILE: &str = "messages.jsonl";
