+++ src-tauri/src/core/threads/mod.rs
@@ -11,6 +11,7 @@
 */
 
 pub mod commands;
+mod constants;
 pub mod helpers;
 pub mod models;
 pub mod utils;
