+++ src-tauri/src/core/threads/utils.rs
@@ -2,11 +2,8 @@ use std::fs;
 use std::path::PathBuf;
 use tauri::Runtime;
 
+use super::constants::{MESSAGES_FILE, THREADS_DIR, THREADS_FILE};
 use crate::core::app::commands::get_jan_data_folder_path;
-
-pub const THREADS_DIR: &str = "threads";
-pub const THREADS_FILE: &str = "thread.json";
-pub const MESSAGES_FILE: &str = "messages.jsonl";
 
 pub fn get_data_dir<R: Runtime>(app_handle: tauri::AppHandle<R>) -> PathBuf {
     get_jan_data_folder_path(app_handle).join(THREADS_DIR)
